+++ internal/file_info.go
@@ -1,0 +1,7 @@
+package internal
+
+type FileInfo struct {
+	Path    string
+	Name    string
+	Version *Version
+}