+++ internal/windows/windows.go
@@ -5,28 +5,21 @@ import (
 	types "github.com/blbrdv/ezstore/internal"
 	"github.com/pterm/pterm"
 	"os/exec"
-	"regexp"
 	"strings"
 )
 
 // Install package if its version higher that installed counterpart.
-func Install(filePath string) error {
-	fileNameRegexp := regexp.MustCompile(`^([0-9a-zA-Z.-]+)[-_]v?([\d\.]+)\.`)
-	pathSlice := strings.Split(filePath, "\\")
-	matches := fileNameRegexp.FindStringSubmatch(pathSlice[len(pathSlice)-1])
-
-	name := matches[1]
-	newVersion, err := types.NewVersion(strings.TrimSuffix(matches[2], "."))
-	if err != nil {
-		return err
-	}
-
-	cmd := exec.Command("powershell", "-Command",
+func Install(file types.FileInfo) error {
+	cmd := exec.Command(
+		"powershell",
+		"-Command",
 		fmt.Sprintf(
 			"Get-AppxPackage -Name %s* | Sort-Object -Property Version | Select-Object -ExpandProperty Version -Last 1",
-			name))
-	result, err := cmd.CombinedOutput()
+			file.Name,
+		),
+	)
 
+	result, err := cmd.CombinedOutput()
 	if err != nil {
 		return err
 	}
@@ -44,13 +37,15 @@ func Install(filePath string) error {
 		}
 	}
 
-	if installedVersion.LessThan(newVersion) {
-		cmd = exec.Command("powershell", "-NoProfile", "Add-AppxPackage", "-Path", filePath)
+	if installedVersion.LessThan(file.Version) {
+		cmd = exec.Command("powershell", "-NoProfile", "Add-AppxPackage", "-Path", file.Path)
 		if err = cmd.Run(); err != nil {
 			return err
 		}
 
-		pterm.Info.Println(fmt.Sprintf("Package %s %s installed", name, newVersion.String()))
+		pterm.Info.Printfln("Package %s %s installed.", file.Name, file.Version.String())
+	} else {
+		pterm.Info.Printfln("Package %s %s already installed. Skipping.", file.Name, installedVersion)
 	}
 
 	return nil